[Refactor] 애니메이션 라이브러리 분리
framer만 사용
