+++ src/shared/hook/useScrollAnimation.ts
@@ -1,24 +1,7 @@
-import { useScroll, useTransform } from "framer-motion";
-
 /**
- * 스크롤 기반 애니메이션 훅
- * @param containerRef 컨테이너 요소의 ref
  * @returns 애니메이션 관련 값들
  */
-export default function useScrollAnimation(
-  containerRef: React.RefObject<HTMLDivElement | null>
-) {
-  // 스크롤 관련 애니메이션을 위한 설정
-  const { scrollYProgress } = useScroll({
-    target: containerRef,
-    offset: ["start end", "end start"],
-  });
-
-  // 스크롤 위치에 따른 애니메이션 값 변환 - 더 넓은 범위와 큰 변화값 적용
-  const opacity = useTransform(scrollYProgress, [0, 0.3], [0, 1]);
-  const y = useTransform(scrollYProgress, [0, 0.5], [70, 0]);
-  const scale = useTransform(scrollYProgress, [0, 0.4], [0.9, 1]);
-
+export default function useScrollAnimation() {
   // 리듀스 모션 설정 확인 (접근성)
   const prefersReducedMotion =
     typeof window !== "undefined"
@@ -41,9 +24,9 @@ export default function useScrollAnimation(
 
   const style = {
     willChange: "transform, opacity", // 성능 최적화
-    opacity: prefersReducedMotion ? 1 : opacity,
-    y: prefersReducedMotion ? 0 : y,
-    scale: prefersReducedMotion ? 1 : scale,
+    opacity: prefersReducedMotion ? 1 : 0,
+    y: prefersReducedMotion ? 0 : 0,
+    scale: prefersReducedMotion ? 1 : 0.92,
   };
 
   return {
